requirements.txt
amqp==5.3.1
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==24.8.0
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
cffi==2.0.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==46.0.4
csp==0.14.0
daphne==4.2.1
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
Django==5.1.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-countries==7.6.1
django-extensions==4.1
django-phonenumber-field==8.0.0
django-post-office==3.9.0
django-timezone-field==7.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
et_xmlfile==2.0.0
greenlet==3.3.1
httpie==3.2.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.1
multidict==6.7.0
mypy-extensions==1.0.0
numpy==2.3.3
openpyxl==3.1.5
packaging==24.1
pandas==2.3.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
phonenumbers==8.13.54
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.2.2
psycopg2-binary==2.9.10
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyotp==2.9.0
pypdfium2==4.30.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
qrcode==8.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.20.0
ruamel.yaml==0.19.1
sendgrid==6.12.5
sentry-sdk==2.42.0
service-identity==24.2.0
setuptools==80.9.0
six==1.16.0
SQLAlchemy==2.0.46
sqlparse==0.5.1
tabula-py==2.10.0
tinycss2==1.2.1
tomli==2.0.2
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==2.1.1
zope.interface==7.2